feat(484): add unit tests and refactor additional skill package name
