--- src/main/java/fr/avenirsesr/portfolio/additional/skill/domain/model/AdditionalSkillsPaginated.java
@@ -1,8 +1,0 @@
-package fr.avenirsesr.portfolio.additional.skill.domain.model;
-
-import fr.avenirsesr.portfolio.shared.domain.model.PageInfo;
-import io.swagger.v3.oas.annotations.media.Schema;
-import java.util.List;
-
-@Schema(requiredProperties = {"data", "page"})
-public record AdditionalSkillsPaginated(List<AdditionalSkill> data, PageInfo page) {}
