--- src/main/java/fr/avenirsesr/portfolio/additional/skill/domain/model/SegmentDetail.java
@@ -1,24 +1,0 @@
-package fr.avenirsesr.portfolio.additional.skill.domain.model;
-
-import lombok.Getter;
-import lombok.Setter;
-
-@Getter
-@Setter
-public class SegmentDetail {
-  private String code;
-  private String libelle;
-
-  private SegmentDetail(String code, String libelle) {
-    this.code = code;
-    this.libelle = libelle;
-  }
-
-  public static SegmentDetail create(String code, String libelle) {
-    return new SegmentDetail(code, libelle);
-  }
-
-  public static SegmentDetail toDomain(String code, String libelle) {
-    return create(code, libelle);
-  }
-}
