--- src/main/java/fr/avenirsesr/portfolio/additional/skill/infrastructure/adapter/mapper/SegmentDetailMapper.java
@@ -1,9 +1,0 @@
-package fr.avenirsesr.portfolio.additional.skill.infrastructure.adapter.mapper;
-
-import fr.avenirsesr.portfolio.additional.skill.domain.model.SegmentDetail;
-
-public interface SegmentDetailMapper {
-  static SegmentDetail toDomain(String code, String libelle) {
-    return SegmentDetail.toDomain(code, libelle);
-  }
-}
