--- src/main/java/fr/avenirsesr/portfolio/additional/skill/infrastructure/adapter/model/DomaineCompetence.java
@@ -1,3 +1,0 @@
-package fr.avenirsesr.portfolio.additional.skill.infrastructure.adapter.model;
-
-public record DomaineCompetence(String code, String libelle, String codeArborescence) {}
