--- src/main/java/fr/avenirsesr/portfolio/additional/skill/infrastructure/adapter/model/Enjeu.java
@@ -1,4 +1,0 @@
-package fr.avenirsesr.portfolio.additional.skill.infrastructure.adapter.model;
-
-public record Enjeu(
-    String code, String libelle, String codeArborescence, DomaineCompetence domaineCompetence) {}
