--- src/main/java/fr/avenirsesr/portfolio/additional/skill/infrastructure/adapter/model/EscoCompetence.java
@@ -1,3 +1,0 @@
-package fr.avenirsesr.portfolio.additional.skill.infrastructure.adapter.model;
-
-public record EscoCompetence(String uri, String libelle) {}
