--- src/main/java/fr/avenirsesr/portfolio/additional/skill/infrastructure/adapter/model/MacroCompetence.java
@@ -1,12 +1,0 @@
-package fr.avenirsesr.portfolio.additional.skill.infrastructure.adapter.model;
-
-public record MacroCompetence(
-    String type,
-    String code,
-    String libelle,
-    String codeOgr,
-    String riasecMajeur,
-    String codeArborescence,
-    String sousCategorie,
-    String maturite,
-    Objectif objectif) {}
