--- src/main/java/fr/avenirsesr/portfolio/additional/skill/infrastructure/adapter/model/Objectif.java
@@ -1,3 +1,0 @@
-package fr.avenirsesr.portfolio.additional.skill.infrastructure.adapter.model;
-
-public record Objectif(String code, String libelle, String codeArborescence, Enjeu enjeu) {}
