+++ src/main/java/fr/avenirsesr/portfolio/additionalskill/application/adapter/mapper/AdditionalSkillMapper.java
@@ -1,0 +1,18 @@
+package fr.avenirsesr.portfolio.additionalskill.application.adapter.mapper;
+
+import fr.avenirsesr.portfolio.additionalskill.application.adapter.dto.AdditionalSkillDTO;
+import fr.avenirsesr.portfolio.additionalskill.domain.model.AdditionalSkill;
+import java.util.List;
+
+public interface AdditionalSkillMapper {
+  static AdditionalSkillDTO toAdditionalSkillDTO(AdditionalSkill additionalSkill) {
+    return new AdditionalSkillDTO(
+        additionalSkill.getId().toString(),
+        additionalSkill.getPathSegments().getSkill().getLibelle(),
+        List.of(
+            additionalSkill.getPathSegments().getIssue().getLibelle(),
+            additionalSkill.getPathSegments().getTarget().getLibelle(),
+            additionalSkill.getPathSegments().getMacroSkill().getLibelle()),
+        additionalSkill.getType().getValue());
+  }
+}
