+++ src/main/java/fr/avenirsesr/portfolio/additionalskill/domain/model/PathSegments.java
@@ -1,0 +1,31 @@
+package fr.avenirsesr.portfolio.additionalskill.domain.model;
+
+import lombok.Getter;
+import lombok.Setter;
+
+@Getter
+@Setter
+public class PathSegments {
+  private SegmentDetail issue;
+  private SegmentDetail target;
+  private SegmentDetail macroSkill;
+  private SegmentDetail skill;
+
+  private PathSegments(
+      SegmentDetail issue, SegmentDetail target, SegmentDetail macroSkill, SegmentDetail skill) {
+    this.issue = issue;
+    this.target = target;
+    this.macroSkill = macroSkill;
+    this.skill = skill;
+  }
+
+  public static PathSegments create(
+      SegmentDetail issue, SegmentDetail target, SegmentDetail macroSkill, SegmentDetail skill) {
+    return new PathSegments(issue, target, macroSkill, skill);
+  }
+
+  public static PathSegments toDomain(
+      SegmentDetail issue, SegmentDetail target, SegmentDetail macroSkill, SegmentDetail skill) {
+    return create(issue, target, macroSkill, skill);
+  }
+}
