+++ src/main/java/fr/avenirsesr/portfolio/additionalskill/domain/port/input/AdditionalSkillService.java
@@ -1,0 +1,7 @@
+package fr.avenirsesr.portfolio.additionalskill.domain.port.input;
+
+import fr.avenirsesr.portfolio.additionalskill.domain.model.AdditionalSkillsPaginated;
+
+public interface AdditionalSkillService {
+  AdditionalSkillsPaginated getAdditionalSkills(Integer page, Integer pageSize);
+}
