+++ src/main/java/fr/avenirsesr/portfolio/additionalskill/domain/port/output/AdditionalSkillCache.java
@@ -1,0 +1,7 @@
+package fr.avenirsesr.portfolio.additionalskill.domain.port.output;
+
+import fr.avenirsesr.portfolio.additionalskill.domain.model.AdditionalSkillsPaginated;
+
+public interface AdditionalSkillCache {
+  AdditionalSkillsPaginated findAll(Integer page, Integer pageSize);
+}
