+++ src/main/java/fr/avenirsesr/portfolio/additionalskill/infrastructure/adapter/mapper/SegmentDetailMapper.java
@@ -1,0 +1,9 @@
+package fr.avenirsesr.portfolio.additionalskill.infrastructure.adapter.mapper;
+
+import fr.avenirsesr.portfolio.additionalskill.domain.model.SegmentDetail;
+
+public interface SegmentDetailMapper {
+  static SegmentDetail toDomain(String code, String libelle) {
+    return SegmentDetail.toDomain(code, libelle);
+  }
+}
