+++ src/main/java/fr/avenirsesr/portfolio/additionalskill/infrastructure/adapter/model/DomaineCompetence.java
@@ -1,0 +1,3 @@
+package fr.avenirsesr.portfolio.additionalskill.infrastructure.adapter.model;
+
+public record DomaineCompetence(String code, String libelle, String codeArborescence) {}
