+++ src/main/java/fr/avenirsesr/portfolio/additionalskill/infrastructure/adapter/model/Enjeu.java
@@ -1,0 +1,4 @@
+package fr.avenirsesr.portfolio.additionalskill.infrastructure.adapter.model;
+
+public record Enjeu(
+    String code, String libelle, String codeArborescence, DomaineCompetence domaineCompetence) {}
