+++ src/main/java/fr/avenirsesr/portfolio/additionalskill/infrastructure/adapter/model/EscoCompetence.java
@@ -1,0 +1,3 @@
+package fr.avenirsesr.portfolio.additionalskill.infrastructure.adapter.model;
+
+public record EscoCompetence(String uri, String libelle) {}
