+++ src/main/java/fr/avenirsesr/portfolio/additionalskill/infrastructure/adapter/model/MacroCompetence.java
@@ -1,0 +1,12 @@
+package fr.avenirsesr.portfolio.additionalskill.infrastructure.adapter.model;
+
+public record MacroCompetence(
+    String type,
+    String code,
+    String libelle,
+    String codeOgr,
+    String riasecMajeur,
+    String codeArborescence,
+    String sousCategorie,
+    String maturite,
+    Objectif objectif) {}
