+++ src/main/java/fr/avenirsesr/portfolio/additionalskill/infrastructure/adapter/model/Objectif.java
@@ -1,0 +1,3 @@
+package fr.avenirsesr.portfolio.additionalskill.infrastructure.adapter.model;
+
+public record Objectif(String code, String libelle, String codeArborescence, Enjeu enjeu) {}
